docker/process-manager/pyproject.toml
[tool.poetry]
name = "process-manager"
version = "0.1.0"
description = ""
authors = ["GIROS-DIT UPM <gi.giros@upm.es>"]

[tool.poetry.dependencies]
python = ">=3.9,<4.0"
fastapi = "^0.74.1"
uvicorn = "^0.17.5"
aiofiles = "^0.8.0"
python-multipart = "^0.0.5"
nipyapi = "^0.19.0"
pydantic = "^1.8.2"
PyYAML = "^5.4.1"
requests = "^2.26.0"
motor = "^3.0.0"

[tool.poetry.dev-dependencies]
pytest = "^5.2"
flake8 = ">=4.0.1"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
